Add helper for common mailer behaviour
