+++ app/mailers/application_mailer.rb
@@ -9,29 +9,18 @@
 # Base class for Mailers
 class ApplicationMailer < ActionMailer::Base
   include FeatureFlagsHelper
+  include ShinyMailerHelper
 
-  before_action :set_view_paths
   before_action :set_site_name
-
-  track open: -> { track_opens? }, click: -> { track_clicks? }
+  before_action :set_view_paths
 
   default from: -> { default_email }
+
+  track open: -> { track_opens? }, click: -> { track_clicks? }
 
   layout 'mailer'
 
   private
-
-  def track_opens?
-    Setting.get( :track_opens )&.downcase == 'yes'
-  end
-
-  def track_clicks?
-    Setting.get( :track_clicks )&.downcase == 'yes'
-  end
-
-  def default_email
-    Setting.get( :default_email ) || ENV[ 'DEFAULT_EMAIL' ]
-  end
 
   def notified_user( email_address, name = nil )
     User.find_by( email: email_address ) ||
@@ -40,13 +29,10 @@ class ApplicationMailer < ActionMailer::Base
   end
 
   def set_site_name
-    @site_name = I18n.t( 'site_name' )
+    @site_name = site_name
   end
 
   def set_view_paths
-    # Add the default templates directory to the top of view_paths
-    prepend_view_path 'app/views/shinycms'
-    # Apply the configured theme, if any, by adding it above the defaults
-    prepend_view_path Theme.current.view_path if Theme.current
+    add_view_paths
   end
 end
