+++ plugins/ShinyForms/app/mailers/shiny_forms/application_mailer.rb
@@ -10,9 +10,10 @@ module ShinyForms
   # Base mailer class - ShinyForms plugin for ShinyCMS
   class ApplicationMailer < ActionMailer::Base
     include FeatureFlagsHelper
+    include ShinyMailerHelper
 
-    before_action :set_view_paths
     before_action :set_site_name
+    before_action :set_view_paths
 
     track open: false, click: false
 
@@ -22,21 +23,12 @@ module ShinyForms
 
     private
 
-    def set_view_paths
-      # Add the default templates directory to the top of view_paths
-      prepend_view_path 'app/views/shinycms'
-      # Add the default templates directory for this plugin above that
-      prepend_view_path 'plugins/ShinyForms/app/views/shiny_forms'
-      # Apply the configured theme, if any, by adding it above the defaults
-      prepend_view_path ::Theme.current.view_path if ::Theme.current
-    end
-
     def set_site_name
-      @site_name = I18n.t( 'site_name' )
+      @site_name = site_name
     end
 
-    def default_email
-      ::Setting.get( :default_email ) || ENV[ 'DEFAULT_EMAIL' ]
+    def set_view_paths
+      add_view_paths( 'plugins/ShinyForms/app/views' )
     end
   end
 end
